\docType{package}
\name{imdfa}
\alias{imdfa}
\alias{imdfa-package}
\alias{package-imdfa}
\title{One-line description}
\description{
  Long description (look at lubridate for an extended
  example)

  Multivariate DFA implementation emphasizing the imaginary
  error component Provides a closed-form solution to the
  filter estimation problem, resulting in
  computationally-fast filter solutions and
  implementations.
}

