\name{print.mdfa}
\alias{print.mdfa}
\title{Print methods for class "mdfa" and "mdfa_coef".}
\usage{
  print.mdfa(x, ...)
}
\arguments{
  \item{object}{fitted mdfa object}
}
\description{
  Prints out the str() representation of the mdfa object.
  For the `mdfa_coef` object, prints the matrix of
  coefficients.
}

