\name{print.predict_mdfa}
\alias{print.predict_mdfa}
\title{Print method for class "predict_mdfa".}
\usage{
  print.predict_mdfa(object)

  print.predict_mdfa(object)
}
\arguments{
  \item{object}{mdfa prediction object}

  \item{object}{mdfa prediction object}
}
\description{
  Long description

  Long description
}

